mv typegen interpretation into its own file
